+++ webui/src/entities/module/types.d.ts
@@ -54,6 +54,7 @@ export type ModuleStatus = {
   hot_allow?: string;
   hide_supported?: string;
   hide_allow?: string;
+  force_bind_capture?: string;
   zn_hide_supported?: string;
   zn_hide_allow?: string;
   zn_hide_zn_module?: string;